addressing PR feedback
